+++ _templates/react/new/index.tsx.ejs.t
@@ -1,0 +1,6 @@
+---
+to: <%= absPath %>/src/index.ts
+---
+import <%= componentName %> from './component/index';
+
+export  {<%= componentName %>};