+++ fp-react/src/modules/data-table/data-table.stories.tsx
@@ -1,0 +1,18 @@
+import React from 'react'
+
+import { ComponentStory, ComponentMeta } from '@storybook/react'
+
+import DataTables from './index'
+
+export default {
+  title: 'Component/DataTables',
+  component: DataTables,
+  layout: 'padded',
+  status: {
+    type: 'alpha',
+  },
+} as ComponentMeta<typeof DataTables>
+
+export const Primary: ComponentStory<typeof DataTables> = () => (
+  <DataTables>Example</DataTables>
+)